Improve project structure.
