+++ Sources/NiftyMarkdownFormatter/Extensions/NSRegularExpression+Ext.swift
@@ -1,0 +1,18 @@
+//
+//  File.swift
+//  
+//
+//  Created by Iiro Alhonen on 7.4.2022.
+//
+
+import Foundation
+
+extension NSRegularExpression {
+    convenience init(_ pattern: String) {
+        do {
+            try self.init(pattern: pattern)
+        } catch {
+            preconditionFailure("Illegal regular expression: \(pattern).")
+        }
+    }
+}
